Modify logic to get PtSettings instance
Since poly toolkit for Unity was made into a Package, this had unforseen
consequences. One of them being the PtSettings.asset file crucial to
Poly at Runtime not being writable when using Poly toolkit for Unity as
a package.

The logic was modified to know copy the PtSettings asset file into the
Assets folder of the end user.
The PtSettings file was moved out of the Resources folder so that it
does not conflict with the final PtSettings asset file location.